floaty good times

+++ Engine/Poo.h
@@ -6,16 +6,16 @@
 class Poo
 {
 public:
-	void Init( int in_x,int in_y,float in_vx,float in_vy );
+	void Init( float in_x,float in_y,float in_vx,float in_vy );
 	void Update();
 	bool TestCollision( const Dude& dude ) const;
 	void Draw( Graphics& gfx ) const;
 private:
-	int x;
-	int y;
+	float x;
+	float y;
 	float vx;
 	float vy;
-	static constexpr int width = 24;
-	static constexpr int height = 24;
+	static constexpr float width = 24.0f;
+	static constexpr float height = 24.0f;
 	bool initialized = false;
 };
